+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
   - Deployment - Keep versions map in deployment module
   - Automated commissioning of NuvlaBox clusters
-  - NuvlaBox Status - expand CPU metrics
+  - NuvlaBox Status - expand CPU metrics and add container-stats attribute
 
 ## [5.12.2] - 2021-04-30